[cmd] Setup config for the stop command
The `stop` command needs the main configuration to know which socket
to connect to.
